chore: rename landing3 to iteam-scale-up
